% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks_width_ggsh}
\alias{breaks_width_ggsh}
\title{Equally spaced breaks for Jalali date and date-time scales}
\usage{
breaks_width_ggsh(width, offset = 0)
}
\arguments{
\item{width}{Interval between consecutive breaks. For details on
supported interval specifications, see "Interval specification".}

\item{offset}{Optional offset or vector of offsets used to shift
breaks away from the default interval boundaries. When multiple
offsets are supplied, they are applied sequentially in the
order given. For details on supported interval specifications,
see "Interval specification".}
}
\value{
A break-generating function that takes a range of \code{jdate} or \code{jdatetime}
values and returns a vector of break positions of the same type.
}
\description{
An equivalent of \code{\link[scales:breaks_width]{scales::breaks_width()}} that returns a break-generating
function for use with \code{ggsh} scales. The returned function generates equally
spaced breaks for a given \code{jdate} or \code{jdatetime} range.
}
\details{
Given a \code{jdate} or \code{jdatetime} range, the returned function:
\enumerate{
\item Extends the range, if necessary, so that it is covered by complete intervals.
\item Generates equally spaced break positions at the specified interval.
\item Applies \code{offset}, if supplied, to shift the resulting break positions.
}
}
\section{Interval specification}{

Intervals are specified as strings of the form \code{"{n} {unit}"}, where "n" is
a positive integer and "unit" is a valid date or time unit.
Examples include "2 days",  "15 min" and "15 minute".
}

