% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demo_jdatetime}
\alias{demo_jdatetime}
\title{Demonstrate scales functions with ggplot2 code}
\usage{
demo_jdatetime(x, ...)
}
\arguments{
\item{x}{A vector of data}
}
\description{
These functions generate ggplot2 code needed to use scales functions for
real code.
}
\keyword{internal}
