% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{fullseq.jdate}
\alias{fullseq.jdate}
\alias{fullseq.jdatetime}
\title{Generate interval-aligned Jalali date and date-time sequences}
\usage{
\method{fullseq}{jdate}(range, size, ...)

\method{fullseq}{jdatetime}(range, size, ...)
}
\value{
An object of the same class as \code{x} (\code{jdate} or \code{jdatetime}).
}
\description{
Methods for \code{\link[scales:fullseq]{scales::fullseq()}} supporting \code{jdate} and \code{jdatetime} vectors.

These methods generate complete sequences of regularly spaced Jalali date or
date-time values covering the range of x. The resulting sequence is aligned
to interval boundaries determined by size, and may extend beyond the
minimum and maximum values of x so that the entire range is covered by
complete intervals.
}
\examples{
fullseq(jdate(c("1405-05-11", "1405-06-15")), "1 month")
fullseq(jdatetime(c("1405-05-11 10:12:00", "1405-05-11 11:45:00")), "1 hour")
}
