% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsh-package.R
\docType{package}
\name{ggsh-package}
\alias{ggsh}
\alias{ggsh-package}
\title{ggsh: Jalali Date/Time Scales for 'ggplot2'}
\description{
A `ggplot2` extension that enables date/time scales based on Jalali calendar. `ggsh` uses classes provided by the `shide` package for mapping aesthetics to date/time types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmollayi/ggsh}
  \item Report bugs at \url{https://github.com/mmollayi/ggsh/issues}
}

}
\author{
\strong{Maintainer}: Mohsen Mollayi \email{mmollayi@gmail.com} [copyright holder]

}
\keyword{internal}
