% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_jdate}
\alias{label_jdate}
\alias{label_jdatetime}
\title{Label Jalali dates and date-times}
\usage{
label_jdate(format = NULL, ...)

label_jdatetime(format = NULL, tz = NULL, ...)
}
\arguments{
\item{format}{A date or date-time format string.}

\item{tz}{Optional time zone for \code{jdatetime} objects.}
}
\value{
A labelling function that takes a vector of \code{jdate} or \code{jdatetime}
values and returns a character vector of the same length.
}
\description{
Create labelling functions that format \code{jdate} and \code{jdatetime} values.
These functions are primarily intended for use with the \code{labels} argument of
\code{ggsh} scales such as \code{scale_x_jdate()} and \code{scale_x_jdatetime()}.
}
\examples{
x <- seq(jdatetime("1405-04-15", format = "\%F"), by = "1 month", length.out = 2)
demo_jdatetime(x)
demo_jdatetime(x, labels = label_jdatetime("\%m/\%d"))
demo_jdatetime(x, labels = label_jdatetime("\%e \%B"))
}
\seealso{
\code{\link[scales:label_date]{scales::label_date()}}
}
