% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_jdate}
\alias{transform_jdate}
\alias{transform_jdatetime}
\title{Transformation for Jalali dates and date-times}
\usage{
transform_jdate()

transform_jdatetime(tz = NULL)
}
\arguments{
\item{tz}{Optional time zone used when transforming \code{jdatetime} values.
If \code{NULL}, the first non-NULL time zone found in the data is used.}
}
\value{
A transformation object created with \code{\link[scales:new_transform]{scales::new_transform()}}.
}
\description{
Create transformation objects for \code{jdate} and \code{jdatetime} values for use
with \code{ggsh} scales such as \code{scale_x_jdate()} and \code{scale_x_jdatetime()}.
}
\examples{
dates <- jdate(c("1400-01-01", "1401-01-01", "1402-01-01"))
# Create a transformation object
trans <- transform_jdate()
# Transform Jalali dates to numeric values
trans$transform(dates)
# Recover the original dates
trans$inverse(trans$transform(dates))
# Generate and format axis breaks
trans$format(trans$breaks(range(dates)))

# Use a specific time zone when converting back from numeric values
trans <- transform_jdatetime(tz = "UTC")
times <- jdatetime("1403-01-01 09:00:00", tz = "Asia/Tehran") + 0:1
trans$inverse(trans$transform(times))
}
\seealso{
\code{\link[scales:new_transform]{scales::new_transform()}}, \code{\link[scales:transform_date]{scales::transform_date()}}, \code{\link[scales:transform_time]{scales::transform_time()}}
}
