% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demo_jdatetime}
\alias{demo_jdatetime}
\title{Demonstrate a jdatetime scale}
\usage{
demo_jdatetime(x, ...)
}
\arguments{
\item{x}{A \code{jdatetime} vector.}
}
\value{
A ggplot object demonstrating the appearance of a jdatetime scale.
}
\description{
Quickly demonstrate the appearance of a jdatetime scale with different
breaks and labels. This is primarily a developer utility for exploring
and testing scale behavior.
}
\examples{
demo_jdatetime(jdate(c("1405-01-01", "1406-01-01")))
}
\seealso{
\code{\link[scales:demo_continuous]{scales::demo_datetime()}}
}
\keyword{internal}
