% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsh-package.R
\docType{package}
\name{ggsh-package}
\alias{ggsh}
\alias{ggsh-package}
\title{ggsh: Jalali Date/Time Scales for 'ggplot2'}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmollayi/ggsh}
  \item Report bugs at \url{https://github.com/mmollayi/ggsh/issues}
}

}
\author{
\strong{Maintainer}: Mohsen Mollayi \email{mmollayi@gmail.com} [copyright holder]

}
\keyword{internal}
